Move replay stable state contract to shared
