+++ apps/web/src/shared/recording-schema/index.ts
@@ -6,3 +6,4 @@ export {
 } from "./validators";
 export { migrateRecordingPackage, migrationRegistry } from "./migrations";
 export { verifyRecordingPackageIntegrity, sha256Blob } from "./integrity";
+export * from "./replayState";
